feat: add schemas for DevOps, Maintainability, and Reliability findings; enhance existing schemas with new fields

+++ services/analysis/app/api/schemas/reports.py
@@ -9,12 +9,12 @@ class ReportResponse(BaseModel):
     repo_id: UUID
     workspace_id: UUID
     branch: str = "main"
-    overall_score: float | None = None
-    performance_score: float | None = None
-    security_score: float | None = None
-    reliability_score: float | None = None
-    maintainability_score: float | None = None
-    devops_score: float | None = None
+    overall_score: int | None = None
+    performance_score: int | None = None
+    security_score: int | None = None
+    reliability_score: int | None = None
+    maintainability_score: int | None = None
+    devops_score: int | None = None
     total_findings: int = 0
     total_files: int = 0
     duration_seconds: int | None = None
